requirements.txt
Django>=5.2.8
djangorestframework>=3.14.0
Pillow>=10.0.0

